增加主题已经主题示例代码

+++ src/main/java/cn/xz/study/proxy/controller/ui/ForwardTableOperatorCellFactory.java
@@ -2,11 +2,18 @@ package cn.xz.study.proxy.controller.ui;
 
 import cn.xz.study.proxy.FxSshService;
 import cn.xz.study.proxy.controller.AbstractController;
+import cn.xz.study.proxy.controller.ProxyFormController;
 import cn.xz.study.proxy.entity.ForwardInfo;
+import cn.xz.study.proxy.util.ButtonUtil;
+import cn.xz.study.proxy.util.DialogUtil;
+import com.jfoenix.controls.JFXButton;
 import javafx.event.ActionEvent;
 import javafx.scene.control.*;
+import javafx.scene.layout.HBox;
 import javafx.util.Callback;
 import lombok.RequiredArgsConstructor;
+
+import java.util.Optional;
 
 /**
  * @author xizhou
@@ -16,18 +23,27 @@ import lombok.RequiredArgsConstructor;
 public class ForwardTableOperatorCellFactory implements Callback<TableColumn<ForwardInfo, Void>, TableCell<ForwardInfo, Void>> {
     private FxSshService sshService = FxSshService.INSTANCE;
     private final AbstractController controller;
+
     @Override
     public TableCell<ForwardInfo, Void> call(TableColumn<ForwardInfo, Void> forwardInfoVoidTableColumn) {
         return new TableCell<ForwardInfo, Void>() {
 
-            private final Button btn = new Button("Action");
-            private final Button updateBtn = new Button("编辑");
-            private final Button delBtn = new Button("删除");
+            private final Button btn = new JFXButton("Action");
+            private final Button updateBtn = new JFXButton("编辑");
+            private final Button delBtn = new JFXButton("删除");
 
             private final ButtonBar buttonBar = new ButtonBar();
+            private final HBox hBox = new HBox();
 
             {
-                buttonBar.getButtons().addAll(btn, updateBtn, delBtn);
+
+                ButtonUtil.primary(btn);
+                ButtonUtil.primary(updateBtn);
+                ButtonUtil.warn(delBtn);
+                hBox.setSpacing(10);
+                hBox.getChildren().addAll(btn, delBtn);
+                //buttonBar.getButtons().addAll(btn, updateBtn, delBtn);
+
                 btn.setOnAction((ActionEvent event) -> {
                     ForwardInfo data = getTableView().getItems().get(getIndex());
                     ForwardInfo finded = sshService.findForwardById(data.getId());
@@ -51,14 +67,21 @@ public class ForwardTableOperatorCellFactory implements Callback<TableColumn<For
                 delBtn.setOnAction(event -> {
                     ForwardInfo data = getTableView().getItems().get(getIndex());
                     try {
+                        Optional<ButtonType> buttonType = new Alert(Alert.AlertType.CONFIRMATION, "确定删除?", ButtonType.OK, ButtonType.CANCEL).showAndWait();
+                        if (buttonType.get().getButtonData().isCancelButton()) {
+                            return;
+                        }
                         if (sshService.deleteForward(data.getId())) {
                             new Alert(Alert.AlertType.INFORMATION, "删除成功", new ButtonType("确定")).show();
+                            //  DialogUtil.showDialog("删除成功");
+                            return;
                         }
                     } catch (Exception e) {
                         e.printStackTrace();
-                        new Alert(Alert.AlertType.INFORMATION, "删除失败", new ButtonType("确定")).show();
+                    } finally {
+                        controller.initializeData();
                     }
-                    controller.initializeData();
+                    new Alert(Alert.AlertType.WARNING, "删除失败", new ButtonType("确定")).show();
                 });
                 updateBtn.setOnAction(event -> {
                     ForwardInfo data = getTableView().getItems().get(getIndex());
@@ -77,10 +100,10 @@ public class ForwardTableOperatorCellFactory implements Callback<TableColumn<For
                     ForwardInfo data = getTableView().getItems().get(getIndex());
                     String name = "开始";
                     if (data.isStarted()) {
-                        name = "暂停";
+                        name = "停止";
                     }
                     btn.setText(name);
-                    setGraphic(buttonBar);
+                    setGraphic(hBox);
                 }
             }
         };
